[minwoo]Fix: requestDto, responseDto 수정

+++ backend/saessak/src/main/java/com/ssafy/saessak/board/dto/BoardRequestDto.java
@@ -15,7 +15,7 @@ public class BoardRequestDto {
     Date boardDate;
     String boardContent;
     String boardTemperature;
-    Integer boardSleepTime;
+    Float boardSleepTime;
     String boardPoopStatus;
     Float boardTall;
     Float boardWeight;
